fix(UserContext): move to its own directory
